rename two files and add fallback for text in CollapsibleAnnotation
